Split auth tests between unit tests and integration tests
